// GenericEditableGrid/generated/ManifestTypes.d.ts
/*
*This is auto generated from the ControlManifest.Input.xml file
*/

// Define IInputs and IOutputs Type. They should match with ControlManifest.
export interface IInputs {
    PrimaryKeyName: ComponentFramework.PropertyTypes.StringProperty;
    PrimaryFieldName: ComponentFramework.PropertyTypes.StringProperty;
    ConfigFileName: ComponentFramework.PropertyTypes.StringProperty;
    EntityName: ComponentFramework.PropertyTypes.StringProperty;
    CSSFileName: ComponentFramework.PropertyTypes.StringProperty;
    ParentLinkedEntityColumnId: ComponentFramework.PropertyTypes.StringProperty;
    RelativeUrl: ComponentFramework.PropertyTypes.StringProperty;
    sampleDataSet: ComponentFramework.PropertyTypes.DataSet;
}
export interface IOutputs {
}